Refactored config router into modular package structure
Split monolithic config.py (~1970 lines) into config/ package with focused modules: schemas.py, helpers.py, settings.py, sources.py, groups.py, and setup.py.
No API changes - all endpoints remain at same paths.
